@@ -1,15 +1,17 @@
 {
-  "name": "botpress-wit",
-  "version": "2.0.11",
-  "description": "This module integrates Wit.ai inside Botpress for NLP",
+  "name": "botpress-recast",
+  "version": "0.0.1",
+  "description": "This module integrates Recast.AI inside Botpress for NLP",
   "main": "bin/node.bundle.js",
   "botpress": {
-    "menuText": "Wit.ai",
-    "menuIcon": "bubble_chart",
+    "menuText": "Recast.AI",
+    "menuIcon": "custom",
     "webBundle": "bin/web.bundle.js"
   },
   "keywords": [
-    "wit",
+    "recast",
+    "recastai",
+    "recast.ai",
     "botpress"
   ],
   "version-manager": {
@@ -19,46 +21,77 @@
     "module-downgrade": "~1.0.x"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "node webpack.js --watch",
-    "compile": "node webpack.js --compile"
+    "test": "npm run lint && npm run mocha",
+    "lint": "eslint ./*.js src/",
+    "mocha": "NODE_ENV=test mocha --compilers js:babel-register --recursive",
+    "watch": "webpack --watch --color",
+    "compile": "webpack --color",
+    "prepare": "rimraf bin && npm run compile"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/botpress/botpress-wit"
+    "url": "https://github.com/hugojosefson/botpress-recast"
   },
-  "homepage": "https://github.com/botpress/botpress-wit",
-  "author": "Botpress Team",
+  "homepage": "https://github.com/hugojosefson/botpress-recast",
+  "author": "Hugo Josefson <hugo@josefson.org>",
   "license": "AGPL-3.0",
   "devDependencies": {
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-rest-spread": "^6.16.0",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-register": "^6.18.0",
+    "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "chai": "^4.1.0",
     "classnames": "^2.2.5",
-    "copy-webpack-plugin": "^4.0.0",
     "core-js": "^2.4.1",
-    "css-loader": "^0.25.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "json-loader": "^0.5.4",
-    "node-sass": "^3.11.2",
-    "react": "^15.3.2",
-    "react-bootstrap": "^0.30.6",
-    "react-dom": "^15.3.2",
-    "react-markdown": "^2.4.2",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.3",
-    "webpack-node-externals": "^1.5.4"
+    "css-loader": "^0.28.4",
+    "eslint": "^4.3.0",
+    "eslint-plugin-react": "^7.1.0",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "file-loader": "^0.11.2",
+    "json-loader": "^0.5.7",
+    "mocha": "^3.5.0",
+    "node-sass": "^4.5.3",
+    "react": "^15.6.1",
+    "react-bootstrap": "^0.31.2",
+    "react-dom": "^15.6.1",
+    "react-markdown": "^2.5.0",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
+    "webpack": "^3.4.1",
+    "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
-    "botpress-version-manager": "^1.0.2",
-    "moment": "^2.17.1",
-    "node-wit": "^4.2.0"
+    "botpress-version-manager": "^1.0.4",
+    "moment": "^2.18.1",
+    "ramda": "^0.24.1",
+    "recastai": "^3.4.1"
+  },
+  "engines": {
+    "node": ">=4.6.0"
+  },
+  "babel": {
+    "env": {
+      "test": {
+        "presets": [
+          [
+            "env",
+            {
+              "targets": {
+                "node": "current"
+              }
+            }
+          ]
+        ],
+        "plugins": [
+          "transform-object-rest-spread",
+          "add-module-exports"
+        ]
+      }
+    }
   }
 }
